refactor: #296 - 노트 임베딩 생성 코드 RAG로 이동
