+++ src/features/ai/rags/note/generate-embedding.ts
